Huge Rework with proper component system for laravel
