+++ app/Http/Livewire/Home/Card.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace App\Http\Livewire\Home;
+
+use Livewire\Component;
+
+class Card extends Component
+{
+    public function render()
+    {
+        return view('livewire.home.card');
+    }
+}
