+++ app/Http/Livewire/Home/Main.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace App\Http\Livewire\Home;
+
+use Livewire\Component;
+
+class Main extends Component
+{
+    public function render()
+    {
+        return view('livewire.home.main');
+    }
+}
