+++ app/Http/Livewire/Home/Modals/EmployeeFiles.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace App\Http\Livewire\Home\Modals;
+
+use Livewire\Component;
+
+class EmployeeFiles extends Component
+{
+    public function render()
+    {
+        return view('livewire.home.modals.employee_files');
+    }
+}
