+++ resources/views/livewire/home/card.blade.php
@@ -1,0 +1,62 @@
+<div class="row">
+    <div class="col-xl-6 col-md-6">
+        <div class="card">
+            <div class="card-block">
+                <div class="row align-items-center m-l-0">
+                    <div class="col-auto">
+                        <i class="icofont icofont-users f-30 text-c-blue"></i>
+                    </div>
+                    <div class="col-auto">
+                        <h6 class="text-muted m-b-10">Happy Customer</h6>
+                        <h2 class="m-b-0">2</h2>
+                    </div>
+                </div>
+            </div>
+        </div>
+    </div>
+    <div class="col-xl-6 col-md-6">
+        <div class="card">
+            <div class="card-block">
+                <div class="row align-items-center m-l-0">
+                    <div class="col-auto">
+                        <i class="icofont icofont-plus f-30 text-c-green"></i>
+                    </div>
+                    <div class="col-auto">
+                        <h6 class="text-muted m-b-10">Happy Customers</h6>
+                        <h2 class="m-b-0">5984</h2>
+                    </div>
+                </div>
+            </div>
+        </div>
+    </div>
+    <div class="col-xl-6 col-md-6">
+        <div class="card">
+            <div class="card-block">
+                <div class="row align-items-center m-l-0">
+                    <div class="col-auto">
+                        <i class="icofont icofont-plus f-30 text-c-red"></i>
+                    </div>
+                    <div class="col-auto">
+                        <h6 class="text-muted m-b-10">Leave</h6>
+                        <h2 class="m-b-0">2</h2>
+                    </div>
+                </div>
+            </div>
+        </div>
+    </div>
+    <div class="col-xl-6 col-md-6">
+        <div class="card">
+            <div class="card-block">
+                <div class="row align-items-center m-l-0">
+                    <div class="col-auto">
+                        <i class="icofont icofont-plus f-30 text-c-blue"></i>
+                    </div>
+                    <div class="col-auto">
+                        <h6 class="text-muted m-b-10">Happy Customers</h6>
+                        <h2 class="m-b-0">5984</h2>
+                    </div>
+                </div>
+            </div>
+        </div>
+    </div>
+</div>
