+++ resources/views/livewire/home/event.blade.php
@@ -1,0 +1,35 @@
+<div class="row">
+    <div class="col-xl-12 col-12">
+        <div class="card">
+            <div class="card-header">
+                <h5>Event Today</h5>
+            </div>
+            <div class="card-block table-border-style">
+                <div class="dataTables">
+                    <table id="EventListToday" class="table table-striped table-bordered nowrap">
+                        <thead >
+                            <tr role="row" style="height: 0px;">
+                                    <th>Full Name</th>
+                                    <th>Event Name</th>
+                                    <th>Event Type</th>
+                                    <th>Date</th>
+                                    <th>Action</th>
+                            </tr>
+                        </thead>
+                        <tbody>
+                            <tr>
+                                <td>Tiger Nixon</td>
+                                <td>System Architect</td>
+                                <td>System Architect</td>
+                                <td>System Architect</td>
+                                <td><a href="" class="icofont icofont-check mr-3"></a>
+                                    <a href="" class="icofont icofont-close mr-3"></a>
+                                </td>
+                            </tr>
+                        </tbody>
+                    </table>
+                </div>
+            </div>
+        </div>
+    </div>
+</div>
