+++ resources/views/livewire/home/modals/employee_files.blade.php
@@ -1,0 +1,34 @@
+<div class="modal fade" id="modalUploadFileEmployee" tabindex="-1" role="dialog" aria-labelledby="myModalLabel"
+    aria-hidden="true">
+    <div class="modal-dialog modal-lg" role="document">
+        <div class="modal-content">
+            <div class="modal-header text-center">
+                <h4 class="modal-title w-100 font-weight-bold">Upload File</h4>
+                <button type="button" class="close" data-dismiss="modal" aria-label="Close">
+                    <span aria-hidden="true">&times;</span>
+                </button>
+            </div>
+            <div class="modal-body mx-3">
+                <div class="md-form mb-3">
+                    <label>Employee Name</label>
+                    <input type="text" class="form-control form-control-sm" readonly>
+                </div>
+                <div class="md-form mb-3">
+                    <label>CV</label>
+                    <input type="file" class="form-control-file border">
+                </div>
+                <div class="md-form mb-3">
+                    <label>KTP</label>
+                    <input type="file" class="form-control-file border">
+                </div>
+                <div class="md-form mb-3">
+                    <label>Certificate</label>
+                    <input type="file" class="form-control-file border">
+                </div>
+            </div>
+            <div class="modal-footer d-flex justify-content-right">
+                <input type="submit" class="btn btn-primary waves-effect waves-light" value="Submit">
+            </div>
+        </div>
+    </div>
+</div>
